WIKI_CEM/src/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
Django==3.1
django-crispy-forms==1.14.0
django-livereload-server==0.4
django-tinymce==3.4.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2022.1
q==2.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
